Moving stan/services/optimization/do_bfgs_optimize.hpp to stan/services/optimize/do_bfgs_optimize.hpp
